requirements.txt
librosa==0.10.1
numpy==1.24.3
pandas==2.0.2
matplotlib==3.7.1
scikit-learn==1.2.2
pydub==0.25.1
groq==0.4.0
spacy==3.7.2
streamlit==1.32.0
praat-parselmouth==0.4.3
fastapi==0.95.2
uvicorn==0.22.0
python-multipart==0.0.6
jinja2==3.1.2
weasyprint==59.0
python-dotenv==1.0.1
seaborn==0.12.2
